server: tests: reduce number of files, all in one tests shell script
