--- examples/server/tests/features/completions.feature
@@ -1,11 +1,0 @@
-Feature: Completion request
-
-  Scenario Outline: run a completion request
-      Given a prompt <prompt>
-      When we request a completion
-      Then tokens are predicted
-
-    Examples: Prompts
-      | prompt                                                         |
-      | I believe the meaning of life is                               |
-      | Write a detailed analogy between mathematics and a lighthouse. |
